requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
Pillow==11.1.0
opencv-python-headless==4.10.0.84
numpy==2.2.3
aiofiles==24.1.0
